Factor out mimetype handling from the notebook and add to console
