+++ src/notebook/widget.ts
@@ -2,10 +2,9 @@
 // Distributed under the terms of the Modified BSD License.
 'use strict';
 
-import * as CodeMirror
-  from 'codemirror';
-
-import 'codemirror/mode/meta';
+import {
+  IKernelLanguageInfo
+} from 'jupyter-js-services';
 
 import {
   RenderMime
@@ -44,6 +43,10 @@ import {
 import {
   IMetadataCursor
 } from '../common/metadata';
+
+import {
+  mimetypeForLangauge
+} from '../common/mimetype';
 
 import {
   EdgeLocation
@@ -290,7 +293,7 @@ class NotebookRenderer extends Widget {
    */
   protected getMimetype(): string {
     let info = this._langInfoCursor.getValue() as nbformat.ILanguageInfoMetadata;
-    return Private.mimetypeForLangauge(info);
+    return mimetypeForLangauge(info as IKernelLanguageInfo);
   }
 
   /**
@@ -574,34 +577,4 @@ namespace Private {
    */
   export
   const activeIndexChangedSignal = new Signal<ActiveNotebook, number>();
-
-  /**
-   * Get the appropriate mimetype for code cell given language info.
-   */
-  export
-  function mimetypeForLangauge(info: nbformat.ILanguageInfoMetadata): string {
-    // Use the codemirror mode if given since some kernels rely on it.
-    let mode = info.codemirror_mode;
-    let mime = 'text/plain';
-    if (mode) {
-      if (typeof mode === 'string') {
-        if (CodeMirror.modes.hasOwnProperty(mode as string)) {
-          mode = CodeMirror.modes[mode as string];
-        } else {
-          mode = CodeMirror.findModeByName(mode as string);
-        }
-      } else if ((mode as CodeMirror.modespec).mime) {
-        // Do nothing.
-      } else if ((mode as CodeMirror.modespec).name) {
-        let name = (mode as CodeMirror.modespec).name;
-        mode = CodeMirror.findModeByName(name);
-      }
-      if (mode) {
-        mime = (mode as CodeMirror.modespec).mime;
-      }
-    } else if (info.mimetype) {
-      mime = info.mimetype;
-    }
-    return mime;
-  }
 }
